pm: policy: Use proper type in the policy function
Use k_ticks_t instead of using int32_t in pm_policy_next_state()
to avoid conversions all around the place.

+++ subsys/pm/pm.c
@@ -206,7 +206,7 @@ bool pm_system_suspend(k_ticks_t ticks)
 	if (!atomic_test_bit(z_cpus_pm_state_forced, id)) {
 		const struct pm_state_info *info;
 
-		info = pm_policy_next_state(id, (int32_t)ticks);
+		info = pm_policy_next_state(id, ticks);
 		if (info != NULL) {
 			z_cpus_pm_state[id] = *info;
 		}
